+++ test/client/jest.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const { jsWithBabel: tsjestPreset } = require('ts-jest/presets');
 
 module.exports = {
-  rootDir: path.join(__dirname, '..'),
+  rootDir: path.join(__dirname, '../..'),
   moduleDirectories: [
     'node_modules',
     path.join(__dirname, 'src'),
@@ -11,21 +11,23 @@ module.exports = {
   testEnvironment: 'jest-environment-jsdom',
   setupFilesAfterEnv: [
     'jest-extended',
-    '<rootDir>/test/setupAfterEnv.js',
+    '<rootDir>/test/client/setupAfterEnv.js',
   ],
   testMatch: [
     '**/__integration_tests__/**/*.spec.[jt]s?(x)',
     '**/__tests__/**/*.spec.[jt]s?(x)',
+    '!**/*.browser.spec.[jt]s?(x)',
   ],
   moduleNameMapper: {
-    '\\.(css)$': '<rootDir>/test/mocks/style.js',
+    '\\.(css)$': '<rootDir>/test/client/mocks/style.js',
     '\\.(png|jpg|jpeg|gif|ttf|eot|svg|ico|webmanifest|xml)$':
-      '<rootDir>/test/mocks/file.js',
+      '<rootDir>/test/client/mocks/file.js',
+    '^&test/(.*)': '<rootDir>/test/$1'
   },
   transform: tsjestPreset.transform,
   snapshotSerializers: [
-    '<rootDir>/test/snapshot-serializers/styled-classes.js',
-    '<rootDir>/test/snapshot-serializers/children-element.js',
+    '<rootDir>/test/client/snapshot-serializers/styled-classes.js',
+    '<rootDir>/test/client/snapshot-serializers/children-element.js',
   ],
   watchPlugins: [
     'jest-watch-typeahead/filename',
@@ -48,6 +50,7 @@ module.exports = {
   ],
   collectCoverageFrom: [
     '<rootDir>/src/**/*.[jt]s?(x)',
+    '!*.spec.[jt]s?(x)',
     '!<rootDir>/src/index.ts',
     '!*.stories.[jt]s?(x)',
     '!<rootDir>/src/@types/**/*'